Creating the quantize_per_tensor and quantize_per_token logic shaders and impl which are linked with the testing framework.

NOTE: Currently the only input types supported are **half** (fp16) and **float** (fp32). The only output types supported are **byte** (uint8), **char** (int8), **short** (int16), **int** (int32).

Differential Revision: [D75959064](https://our.internmc.facebook.com/intern/diff/D75959064/)

[ghstack-poisoned]

+++ backends/vulkan/test/op_tests/test_utils.cpp
@@ -94,7 +94,8 @@ vkcompute::vkapi::ScalarType from_at_scalartype(c10::ScalarType at_scalartype) {
     case c10::kInt:
       return vkapi::kInt;
     case c10::kLong:
-      return vkapi::kLong;
+      // No support for 64-bit integers
+      return vkapi::kInt;
     case c10::kChar:
       return vkapi::kChar;
     case c10::kByte:
